FUSETOOLS-2768 - logging exception - moved test to an integration test bundle
