+++ syndesis/tests/org.fusesource.ide.syndesis.extension.core.tests.integration/META-INF/MANIFEST.MF
@@ -1,0 +1,12 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %Bundle-Name
+Bundle-SymbolicName: org.fusesource.ide.syndesis.extension.core.tests.integration
+Bundle-Version: 10.3.0.qualifier
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Require-Bundle: org.junit;bundle-version="4.12.0",
+ org.assertj.core;bundle-version="2.1.0",
+ org.fusesource.ide.syndesis.extension.core;bundle-version="10.3.0",
+ org.fusesource.ide.syndesis.extension.ui;bundle-version="10.3.0"
+Bundle-Vendor: %Bundle-Vendor
+Bundle-Activator: org.fusesource.ide.syndesis.extension.core.tests.integration.SyndesisExtensionIntegrationTestsActivator
