+++ syndesis/tests/org.fusesource.ide.syndesis.extension.core.tests.integration/OSGI-INF/l10n/bundle.properties
@@ -1,3 +1,3 @@
 #Properties file for org.fusesource.ide.syndesis.extension.ui.tests.integration
 Bundle-Vendor = JBoss by Red Hat
-Bundle-Name = Fuse Tooling Syndesis Extensions UI Test Integration
+Bundle-Name = Fuse Tooling Syndesis Extensions Core Test Integration
