+++ syndesis/tests/org.fusesource.ide.syndesis.extension.core.tests.integration/build.properties
@@ -1,0 +1,8 @@
+source.. = src/main/java/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               OSGI-INF/l10n/bundle.properties,\
+               OSGI-INF/
+src.includes = *
+src.excludes = src/
