+++ syndesis/tests/org.fusesource.ide.syndesis.extension.core.tests.integration/infinitest.filters
@@ -1,0 +1,1 @@
+.*IT
