+++ syndesis/tests/org.fusesource.ide.syndesis.extension.core.tests.integration/pom.xml
@@ -10,7 +10,7 @@
     <relativePath>..</relativePath>
   </parent>
   
-  <artifactId>org.fusesource.ide.syndesis.extension.ui.tests.integration</artifactId>
+  <artifactId>org.fusesource.ide.syndesis.extension.core.tests.integration</artifactId>
   <packaging>eclipse-test-plugin</packaging>
-  <name>JBoss Fuse Tooling :: Syndesis Support :: Tests :: Syndesis Extensions UI Integration Tests</name>
+  <name>JBoss Fuse Tooling :: Syndesis Support :: Tests :: Syndesis Extensions Core Integration Tests</name>
 </project>
